requirements.txt
fastapi
uvicorn[standard]
pillow
opencv-python
torch        # ou tensorflow
sqlalchemy
databases    # si tu veux un accès asynchrone
python-multipart  # pour l’upload d’images via FastAPI
torchvision

